[feature] Crithon API 서버 및 플러그인 2차 개발

+++ server/src/main/java/org/crichton/domain/dtos/report/InjectionTestDefectDto.java
@@ -14,9 +14,9 @@ public class InjectionTestDefectDto {
 
     // 결함 주입 스펙상의 id
     @JsonProperty("defect_id")
-    private Long defectId;
+    private int defectId;
 
     // 안정 판정 스펙상의 id
     @JsonProperty("violation_id")
-    private Long violationId;
+    private int violationId;
 }
